frontend: log pending build records only when debugging
Today we almost ran out of disk space on copr-frontend because we are
mass-rebuilding rubygems, and periodically logging all pending builds
resulted in 20G `copr-frontend.log`.

I think we can simply log the number of pending builds instead of
their whole records. That might be useful for debugging though, so I
am keeping the original command but switching it from `log.info` to
`log.debug`.