cli: add regenerate-repos command
Fix #1915
See PR#1907

In the PR#1907, @tstellar already implemented the support for
regenerating repositories in APIv3 (on the frontend side) and in the
python-copr package. I am just finishing it by adding support for
copr-cli and covering it with beaker test.

Also, there was two typos in python-copr in the API endpoint URL, so I
am fixing them as well.