frontend: support searching by attributes and improve searching overall
It starts to look that we are going to put search bar to the center of
the homepage, see #1890. The design mockup suggests that it should be
available to search by using specific attributes such as project name,
owner name, or package name. I think it is important otherwise the
searching returns too many _false-positives_.

I rewrote the whole `get_multiple_fulltext` function so it may not be
clear on the first sight but it is possible to use searching exactly
as before. Searching the following strings return the exact same
results as previously:

- frostyx
- @copr
- @copr/copr
- frostyx/tracer
- copr-cli
- 2679386

But on top of that, it now recognizes the following GET parameters -
`ownername`, `projectname`, `packagename`, and their combinations.
Also, I didn't do any measuring but the searching feels a bit faster
now.

We didn't have any RFE for this so I wasn't sure what way of
specifying attributes we preferred, so I chose the one requiring the
least magic possible. But if we want to implement some custom
searching language, I think this PR would be a good step towards it.