openid,login by email: guide user, don't do infinite loop
The main problem is that FAS/ipsilon sets the email used during login as
openid.claimed_id instead of actual FAS username. This breaks openid
client logic badly.

flask_openid in this situation redirects to /login/, again, which causes
infinite request loop.

In this commit we create a new decorator to wrap our login method to
detect this situation and prevent it:

0. detect this specific use case

1. drop the `openid_error` from flask session

2. flash a message they hit an auth bug

3. ask them to log out and log in again with FAS username

4. redirect to "/"

Signed-off-by: Tomas Tomecek <ttomecek@redhat.com>