frontend: assure error_handler error is 500
It happened before (#1957) that SQLAlchemy's NoResultFound bubbled up to
our error handler, and the 'getattr(error, "code", 500)' was returning
None (as code was explicitly set to None).  In such situation, instead
of returning a "nice" error 500, it raised another exception in the
error handler leading to much less nice error on page 200.