cli: more detailed "Request is not in JSON format" error
When there's a server-side traceback upon cli API call, the Frontend
logic returns some html-based error page which may contain some useful
error message.  At least when we are debugging.

So newly, when cli is run with --debug option, try to parse the web page
(try beautifulsoup lib, if available) and dump it on stderr.

Merges: #1966