frontend: large project modification timeout fix
On @rubygems/rubygems in Fedora Copr, we faced timeouts when we tried to
disable mock chroots with too many BuildChroots.  Turns out ORM lazy
loaded all the BuildChroot instances for disabled CoprChroot (which in
our case was more than 100k items).

Instead of this, prepare a custom query for related BuildChroots that
can be used with limit() safely.

For the purpose of the query, we need to use FINISHED_STATUSES variable
which was improperly using states, not statuses.  So I'm fixing this
(status is the number, state is the string human-readable alternative).

Fixes: #1925
Closes: #1968