backend, frontend, rpmbuild: fixup ACR handling
When auto createrepo is set OFF, we need to run the initial createrepo
also in devel/ repo, otherwise the builds will fail because we only
execute createrepo_c _after_ every successful build, not _before_ every
build.

The same for changing ACR ON => OFF; at least for the first time this
change is done, createrepo needs to be done in devel/ repo, too.

This change allows us to finally drop the very dangerous
"skip_if_unavailable=1" option in the rpmbuild code.

Update the "when we run creatrepo" documentation table while on it.

Fixes: #1145