frontend: ACR toggle - handle NO_VALUE specially too
It appears that, with older SQLAlchemy implementations (Fedora <= 34),
the initial value in the event listener is sometimes set to NO_VALUE
symbol.  We want to treat NO_VALUE symbol the same way we treat the
NEVER_SET symbol.

The previous commit quite accidentally caused that some of the existing
test cases begun failing.  The reason was the unhandled NO_VALUE symbol.
But issue fixed by this commit might well be the reason why we were
seeing the random missing initial createrepo_c actions (and thus
following failing builds).  We've been advising to execute the
createrepo manually as a work-around till now.

Relates: #1969, #1475, #1192, #833