releng, doc: add helper script to wait until packages are in infra tags
It's not fun to obsessively reload the repository page and manually
search the packages there. I am adding a helper script for that.

https://kojipkgs.fedoraproject.org/repos-dist/f33-infra-stg/latest/x86_64/kojipkgs
https://kojipkgs.fedoraproject.org/repos-dist/f33-infra/latest/x86_64/kojipkgs

During the last release I forgot to update the "f33" to "f34" (for the
builder packages) in the URL and it took me too long to realize
that. With this script, such slip-up shouldn't happen since you simply
take the output of `koji-infratag-staging` (and
`koji-infratag-move-prod`) and pass it to the
`koji-infratag-available` script. Although not yet, since
`koji-infratag-staging` works with the packages in the following
format `copr-rpmbuild-0.53-1.fc34` and this script works with
`copr-rpmbuild-0.53-1.fc34.x86_64.rpm` in order to calculate the URL
based on the pacakge arch. We should unify the script inputs.