distgit: fix import destination
After upgrade to F35, we started seeing the following errors in the
buider-live.log:

    Traceback (most recent call last):
    File "/usr/bin/copr-distgit-client", line 10, in <module>
	main()
    File "/usr/lib/python3.9/site-packages/copr_distgit_client.py", line 371, in main
	sources(args, config)
    File "/usr/lib/python3.9/site-packages/copr_distgit_client.py", line 236, in sources
	download_file_and_check(url_file, kwargs, distgit_config)
    File "/usr/lib/python3.9/site-packages/copr_distgit_client.py", line 130, in download_file_and_check
	download(url, filename)
    File "/usr/lib/python3.9/site-packages/copr_distgit_client.py", line 109, in download
	raise RuntimeError("Can't download file {0}".format(filename))
    RuntimeError: Can't download file copr-cli-1.97.tar.gz

It happens because Copr DistGit started uploading packages into the
wrong lookaside path, e.g.

    /var/lib/dist-git/cache/lookaside/pkgs/copr-cli/copr-cli-1.97.tar.gz/

Notice the missing ownername. We configure the ownername via
`commands.module_name` but it was deprecated and stopped working since
python3-rpkg-1.63. We need to use `commands.repo_name` now:

        # Deprecates self.module_name
        self._repo_name = None

We also need python3-rpkg at least in 1.63-5 because of
https://pagure.io/rpkg/issue/588