frontend: drop duplicit "group" table join
Fix #2003

The traceback was caused by

    sqlalchemy.exc.ProgrammingError: (psycopg2.errors.DuplicateAlias)
    table name "group" specified more than once

That's because we already join the `models.Group` in the
`CoprsLogic.get_multiple`.

Also, not exactly #2003 but a related bug can be reproduced with

    /api_2/projects?group=copr&search_query=copr

It doesn't traceback but it returns results based on the
`search_query` but not limited to the specified `group`. We should
pass `group` and `owner` to the `CoprsLogic.get_multiple_fulltext`.