rpmbuild: better PID for copr-rpmbuild-log
The problem is that the daemonized 'copr-rpmbuild' process EXITs a bit
earlier than it's child process 'copr-rpmbuild-loggify' (which reads the
parent's stdout and stderr through PIPE, and continuously writes that
content to 'main.log').

The 'copr-rpmbuild-log' then reads content of 'main.log' file and puts
the content to it's stdout (which is read by copr-backend over SSH).
The thing is that 'copr-rpmbuild-log' can not quit earlier than the
process which writes 'main.log'.

Therefore, due to that race, 'copr-rpmbuild-log' needs to wait for the
'copr-rpmbuild-loggify' process, not the 'copr-rpmbuild' process.
Otherwise we might see incomplete logs (which often happened before
while running the runtest-makesrpm-fail.sh test-case).

Merges: #2018