frontend: use official EPEL log for chroot icon
We are moving to RHEL+EPEL fore epel-8+ chroots, while staying with
CentOS+EPEL for epel-7 chroots (i.e. not a CentOS only in Copr anymore).
Let's use the official icon instead.

While on it, replace the "Epel for CentOS" string with just EPEL.  EPEL
is "for" all Enterprise Linux variants, not just CentOS.  We could say
"built on/against CentOS", though this will not work either soon.

Merges: #2031