backend, frontend: basic build tagging
We don't need a per-build tagging mechanism ATM, but we need
per-mock-chroot tag propagated to builds, as we have two kinds of the
ppc64le builders.  Power8 machines (might be used for building Fedora
and older Enterprise Linux chroots), and Power9 (can be used for
Enterprise Linux 9+, but also other/older chroots).

The point here is to never assign Power8 machine for the EL 9 chroot
build task.  For this, we need to have a special tag (say Power9, or
Power9_needed) and a correspondingly tagged Resalloc resources.

Also, while we have separate Resalloc pools for the Power8 and Power9
architectures, we need to be able to appropriately set the limits for
each of those; therefore there's the new `builds_max_workers_tag`
option.

Merges: #2032