modules: use dist-git method for builds by default
The custom clone URL is somewhat problematic.  We used to have
'/etc/copr-rpmbuild/main.ini' configured so rpkg-util could use several
DistGit instances, but that did not really cover lot of DistGit
instances.

Now, with rpkg-util v3+ we'd have to completely rebuild the main.ini
config.  There are issues like [1], etc.

This commit causes that we start using the DistGit method for all the
modular builds instead of SCM method.  As long as the specified
component repository is correctly configured on builder side - this will
work.

[1] https://pagure.io/rpkg-util/pull-request/38

Merges: #2055