rpmbuild: copr-distgit-client: add centos-stream configuration
Because GitLab is not a dedicated DistGit forge, in general we might
end-up with multiple DistGits hosted on that forge.  Therefore I'm
adding the 'path_prefixes' configuration option which takes list of
strings of possible clone_url prefixes.  Can look like:

    path_prefixes =
        redhat/centos-stream/rpms
        ...

The (CentOS Stream related, in this case) files will be downloaded from
the correct lookaside cache only if both the hostname and one of the
specified prefixes matches the clone URL.  This leaves us space to
specify other DistGit instances hosted on GitLab in the future (with the
same hostname, but different path prefix).

Fixes: #2015
Merges: #2058