frontend: don't use redis as a middleman when updating hitcounter stats
I am not removing everything redis-related, that will be done in a
follow-up commit. I've also put a lot of TODOs in this commit, I
didn't want to break the compatibility with our already stored
data. We should do it but not as a part of this already big change.