backend: equivalent priority for concurrent sandboxes
With the recent @copr/PyPI tasks (tens of thousands builds) we observed
that other builds in @copr/ namespace aren't processed.  That's because
the WorkerManager.run(timeout=) isn't able to go through that huge bunch
of tasks in the priority queue -- and because the other (submitted
later) had lower priority.

Resolves: #2079
Merges: #2080