doc: document that `branch-fedora` can be done earlier
This approach was tested while F36 was branched, and the process seems
safe.  We had F36 chroot enabled and working in Fedora Copr sooner than
branching was eventually done in the official Fedora repos (even though
fedora-36 was temporarily equivalent to fedora-rawhide).

While on it, I removed the CHROOT_NAME_RELEASE_ALIAS as this hack
shouldn't be needed.