frontend: add /pending/all/ stats
While we still keep the old /pending/ page, this new gives us more
detailed overview for the whole queue (what owners, chroots or projects
filled the queue, numbers of background jobs).

The new route works with just "pending" jobs (contrary to the
'for_backend' route(s)) while — for the performance reasons — it only
needs a limited set of object data (joinedload && load_only tricks are
useable).  This is why this commit defines another data_type=overview.

The new route is pretty expensive, especially for queue of size in tens
of thousands tasks or more.  This is why the page is cached.