backend: copr_fix_gpg: properly regenerate DNF metadata
We can not just run "createrepo_c --skip-stat" from copr_fix_gpg.py
as we otherwise do by default because in copr_fix_gpg.py we are changing
the RPM GPG signatures (and thus the overall package checksum) – while
the RPM filename is not changed.  With The "--skip-stat" option then,
createrepo_c is not able to detect this file change and thus the repo
data stay untouched (and outdated).

Fixes: #2106
Merges: #2119