frontend: make the /pending-jobs/ a streamed site
Use the same approach that is already used for the monitor page.  This
makes the route much less memory demanding.

I added a tweak with "batching" the items in the response, which
minimizes the Flask iterating overhead and the response is at least as
fast as the previous approach.  This actually speeds up the monitor
route as well:

    time for i in 1 2 3 4 5; do
      curl "http://localhost:55555//api_3/monitor?ownername=@rubygems&projectname=rubygems" >/tmp/log$i
    done

Decreases the time from 1m34.575s to 1m25.857s.  The items in monitor
route are rather large (compared to the pending-jobs) so the effect is
not that large.