backend: manager: properly track running workers
The _tracked_workers is now kept precisely in sync with the Redis DB
state.  Previously we cleaned the _tracked_workers set for each
WorkerManager.run(), and it got de-synced.

Previously we cleared the _tracked_workers set() for each run().  This
caused that we started a brand new Worker for the same (already running)
task in the queue in every run().  This was a fork bomb.

Now, when we do not clear the _tracked_workers set  we have to re-work a
bit the _start_tracking_worker() method.  It had two purposes (calculate
limits, and add the worker to _tracked_workers).  Now we have those two
purposes separated, so the method is renamed.

We also have to load the initial state of _tracked_workers from RedisDB,
otherwise we are unable to adopt the workers started by previous
dispatcher process (before systemctl restart copr-backend.target).