frontend: detect ClientDisconnected errors
The ClientDisconnected is just HttpException sub-class, so we already
handle that.  It is though good to special-case it and properly dump
what is happening to logs.

I realized that previous commit stopped logging the HTTP errors to logs
entirely (we only have error code in the access_log).  It is probably
better to say explicit, and still have a single-line error for each
non-200 error we send to users.

Relates: #1600