backend: ignore 404 hits
That provides validation that such project and chroot either exists or
at least existed in the past. I think we want to exclude 404 hits and
count everything else. We could count only 200 but there might be
redirects or temporarily broken server returning 500 and I think we
want to count those.

Merges: #2189