rpmbuild: define copr-specific macros also for SRPM builds
Fix #2094
Merges: #2197

I am also trying to unify the macros for all methods. Ideally, we
would run every provider within Mock and have just one way of defining
macros but that's not a small change.

Instead, I picked all the custom macros and moved them to one
function. It is still up to the providers/builders how the macros will
be defined (using --define, generating ~/.rpmmacros, etc) but the list
of the macros and their values are always the same.