frontend: de-duplicate Packages for each CoprDirectory
The Package model served two purposes, (a) it was an aggregation
mechanism for project build-state analysis routes like "monitor" (the
Build -> Package -> CoprDir relation), and (b) it stored "build recipes"
for the actual building.

The (a) part was a clear redundancy though, because Build is directly
linked to CoprDir so we can actually get this info a bit cheaper way.
The (b) was meant to be a "flexible" configuration addition for the
future, but we eventually never implemented "custom build recipes" per
CoprDir; neither API nor Web-UI is not designed for this.

If we ever wanted multiple recipes per Package in the future, we should
instead go with a separate model like "BuildRecipe -> Package", and
avoid restricting each of the BuildRecipes to a concrete CoprDir
(CoprDir model is only an implementation detail for categorizing
_builds_ within project, not packages).

Fixes: #617
Fixes: #2082
Relates: #2175