rpmbuild: undefine %dist for SRPM builds
It was reported several times, that the %dist part of the version is
confusing because it is evaluated to the Fedora versions that is
currently running on our builders, instead of the target chroot.

We could either add some tooltip explaining this, or not put the %dist
value into the SRPM name in the first place.

Merges: #2208