cli, python: simplify cli configuration
There's no need to setup token defaults both in copr-cli and
python-copr.  Simply delegate the default "Fedora Copr" configuration to
the python-copr logic.

Make sure that we fail verbosely if user explicitly specifies
a non-existing token file.

Drop the complicated error message that we stopped printing out anyway
(at least not in appropriate situations for which that message was
originally designed).

Fixes: #2253