backend: add option for infinite number of attempts to the hitcounter script
I don't think this will be useful in production for two reasons. We
don't want to use this option for cron tasks. Also, in production,
once we process an access file, we delete it from S3. Therefore
re-running the script won't count anything twice nor waste any time
with already counted files.

But this option is useful when debugging. I don't want the script to
die after hours of counting accesses that would be counted again when
re-running the script.