frontend: pass a list of projects instead of a query
Our logic in `templates/coprs/show/all.html` is to show pinned
projects if there are any. If not, show a user's recent projects if
there are any. The condition was always true though because we passed
a `flask_sqlalchemy.BaseQuery` object to the template instead of the
list of projects.

Merges: #2366