backend, common, dist-git: correct background worker start
By always calling the self._wm_started() method.  This needs to be done
in the parent BackgroundWorker class so we don't forget about it in
inheriting classes -> and it needs to be done ASAP.  To keep the order
of things to be done (first call _wm_started(), then redirect logs to
redis for the backend side, and then handle_task()) - there's now a new
method preparations_for_manager() which is called after _wm_started()
but before handle_task().