rpmbuild: decode URL encoded characters such as %5E into caret
There is a valid use-case for packages with a caret symbol ^ in their
version. See the packaging guidelines for snapshots.
https://docs.fedoraproject.org/en-US/packaging-guidelines/Versioning/#_snapshots
This applies for other characters as well (tilde, dot) but let's focus
on the caret.

If you submit such a package via URL like this

https://example.com/python-nss-1.0.1^20210803hg9de14a6f77e2-4.fc38.src.rpm

everything works as expected. However, if the caret is URL encoded

https://example.com/python-nss-1.0.1%5E20210803hg9de14a6f77e2-4.fc38.src.rpm

then we download the file and keep the URL encodeding in it. First,
this is IMHO wrong, and second it causes issues on multiple places
down the line (logging, importing on DistGit, ...)

I am decoding the URL filename.