common: periodically dump the priority queue to a JSON file
Many times in the past I wanted to see how exactly the priority queue
looks like at a given time but we didn't have a way of doing it. I
wanted to implement some script for displaying it, but that's not
possible because we AFAIK keep all the priority queue state in a
variable (which of course is not accessible from the outside). So what
about periodically dumping the queue into a JSON file by the
`Dispatcher` itself?

I think this would be good enough for all the times that I needed it
in the past but in case it was not sufficient, this would allow us to
implement any kind of script to visualize the queue any way we want.