frontend: allow uploading artificial payload in custom webhook
It turns out non-json payload was never actually handed-over to a custom
script.  This was caused by the artificial request.json check in the
HookContentStorage.__init__ method.  But with the previous Flask
versions we were at at least able to successfully submit the custom
webhook build even with an empty payload without specified content type.

After the migration to F37 with new Flask the check for request.json
started raising the exception 400 for empty payloads (causing #2416):

    Did not attempt to load JSON data because the request Content-Type
    was not 'application/json'

The JSON check is now replaced with request.get_data() which is a safe
call in any case (form submitted, json data, uploaded file).

Fixes: #2416