frontend: allow automation (cron) to remove builds
To work-around the missing abstraction in delete_build() logic, we
previously "faked" the user argument with the actual build.copr.owner,
with believe that it will never fail.

That `copr.owner` is actually not guaranteed to have build/admin access
to corresponding 'copr'.  Not after a4e7a5aa467cad6132eb9d00170.

Let's invent AutomationUser (having a common ancestor with User)
providing the necessary API to perform permission checks for build
removals, and grant the AutomationUser instances the necessary
permissions.

While on it, tweak the logging and error reporting a bit so there's is
more obvious what is going on.

Fixes: #2523
Fixes: #2019