dist-git: skip test_no_op_1
new version of python-rpkg introduced unexpected behavior of commit
where it commits empty commit by default. This causes a bug if multiple
importing of a package happens in copr and the latter import of a
package is without any change, the following `commit` will make a new
commit hash because it allows empty commits by default.

I created an issue about it in rpkg. See
https://pagure.io/rpkg/issue/677

in the meantime the temporary solution of this is to skip the
test and remove the --allow-empty from rpkg directly in the code
Once this issue is resolved unskip the test.