frontend: for-DNF5 API rpmrepo route
The new route provides all the necessary info needed to let the rest of
the "repofile" creation task on DNF5 Copr plugin.  The route lists all
the available CoprDirs, CoprChroots (even soon to be deleted) and
dependencies.

The route contains seemingly useless parts like <dirname> and
<name_release>.  These are only needed to keep the repo downloading
statistics working with DNF5.

Closes: #2603