dist-git: refresh pyrpkg.Commands after import_srpm
in case of importing, the content of directory in `reponame`
changes. To update the state of `Commands` class isn't an easy
process - look how much logic pyrpkg.cli.cliClient.load_cmd has
and the logic is written for context in cliClient, not for Commands
which we are using. Considering this wouldn't be an easy task for
pyrpkg to implement, we can refresh the Commands class after every
import - even if it is not elegant