beaker-tests-sanity: don't test project searching
See #2578

It is now not possible to update whoosh index on demand because the
route timeouts. Instead, we periodically rebuild the whoosh index via
Cron. It was a compromise between this, and having the index too large
and causing OOM issues.

Unfortunately, it is not possible to easily test searching now. We
need to upgrade from whoosh to a real search engine.