beaker-tests-sanity: don't test project searching pt.2
See PR #2731

I made a mistake and forgot to comment the following code as well:

    cleanProject "${NAME_PREFIX}Project8"
    cleanProject "${NAME_PREFIX}Project9"

Now the project deletion fails because the projects are not being
created in the first place. This PR fixes it. Alternatively I could
comment the project deletion but then we will surely forget to
uncoment them in the future because they are too far from the project
search tests.