backend: don't enforce appstream-builder in copr_prune_results
We recently disabled appstream metadata for all projects, because the
appstream-builder is simply too I/O demanding:
https://github.com/fedora-copr/copr/issues/2577

We though forgot that the appstream checkbox is not reflected in the
copr_prune_results.py script at all.  Newly it is reflected, meaning
that the overall copr_prune_results.py run will be much faster.

Closes: #2749