dist-git: fix test_do_import test
The 'has_calls" actually never worked - it was a typo which had a no-op
effect - and with the Python 3.12 in Fedora 39+ it finally started
failing.

Let's switch to a valid 'assert_has_calls' and update the expected
output (which is a bit artificial after the years, but it doesn't
matter).

Relates: #2811