beaker: align tests with the August 2023 release
Mock 5.0 is enforced by the new Copr release, and there's the
bootstrap_image feature that can not use "fedora image" for EPEL target
chroot

The unspecified `Epoch` tag in spec file is newly evaluated as "null"
value in the BuildChrootResults table, and returned as None to end-user
(null JSON value).

The default PyPI method is pyp2spec, not pyp2rpm now.  Let's use a
package that is easily buildable with pyp2spec (motionpaint doesn't
comply with the new licensing format, we could submit PR but that would
require a release into PyPI, and ...).

Skip one test related to a ExclusiveArch parsing for now; our ppc64le
builders are somewhat broken or non-standard, see #2870.

Several cleanup actions are now wrapped with cleanupAction, so we can
keep the intermediate files/projects hanging around with
COPR_CLEANUP=false when debugging testsuite failures.