beaker: predictable sub-package ExclusiveArch test
The package was crafted to properly match the use-case we want to test,
two chroots skipped, two processed.  The sub-package ExclusiveArch is
even more filtered (x86_64 only), but it doesn't affect the overall
build on aarch64.  Long term, we should stop shipping the subpackage
into aarch64, though (as is done in Fedora Pungi).