backend: react on staled SSH connections
All the SSHConnection "public" methods now have subprocess_timeout
option.  If the option is set, and the local utility (ssh or rsync)
doesn't finish within the specified timeout, SSHConnectionError is
raised as in any other connection-issues situation.

The SSHConnection.run() and SSHConnection.run_expensive() methods have
the default timeout set to 180s by default.  The reason is that both of
those commands are always expected to finish reasonably quick (much less
than 10s).  If not, after 180s period we can either expect the builder
is dead, or optionally retry.

The timeout detection should help us to cover weird situations like in
the issue #2888.  If the remote process is in 'D' state (uninterruptible
sleep), we simply kill the ssh client (subprocess) locally.  Subsequent
SSHConnectionError handlers are able to close the corresponding Resalloc
ticket, and take a new builder.

Relates: #2888