frontend: unify naming convention for safe methods
Originally, safe methods were considered as methods raising ObjectNotFound
for API (mainly with session.one() method). Now methods which returns None
or ignores the error completely are also considered as "safe".

For the sake of consistency:
- *_or_none -> returns None if object was not found in the database
- *_safe -> methods that ignores error completely (they basically
   log and/or are safe from exceptions)

and renaming some of the current safe methods to normal ones that aren't
safe by the new rules