backend: longer timeout for fallback generating GPG keys after build
Fix #2911
Fix #2942

When frontend is creating a new project, it creates an action to
generate GPG keys. This may fail because of a temporary keygen
downtime or network issues.

Once a SRPM is built, we try to use the GPG key and it doesn't exist,
there is a fallback, trying to generate the key again. We can try
longer.

To trigger this fallback, you can do:

    docker-compose stop keygen-httpd
    copr create test-srpm-key --chroot fedora-38-x86_64
    copr buildscm --clone-url ... test-srpm-key
    # Wait few minutes until the action fails
    docker-compose start keygen-httpd