frontend: Add backref relationship between Package and Build
... to silence sqlalchemy warnings

The Package.builds column is used only for read only purposes
and needs to be ordered accordingly against Build.id. Removing
viewonly=True parameter and making it read-write column will
result in errors since the relationship is used only for loading
objects and not for any persistence operation.

Setting the relationship to read-write would result in errors
since the relationship.sync_backref flag is set to True and
sqlalchemy tries then to sync the Package.builds column which
was never commited - thus "This session is in 'committed' state;
no further SQL can be emitted within this transaction." will appear.