backend, dist-git: rename dispatcher scripts
A hurdle when trying to stand-up a simple single-machine Copr
deployment is that both the copr-backend and copr-dist-git packages
install a script to /usr/bin/copr-run-dispatcher, preventing the user
from asking that both be installed on the same system with dnf.

This commit renames the script from backend and dist-git to
copr-run-dispatcher-backend and copr-run-dispatcher-dist-git
respectively.

Fixes #3007.