frontend/oidc: add OIDC_USERNAME_CLAIM
The current OpenID Connect code assumes the existence of a
non-standard field in the UserInfo returned by the provider, limiting
the ability to use non-Ipsilon OIDC providers; however, the
standard forbids using corresponding standardised field
(preferred_username) as a unique user identifier, a property on which
Copr relies.

This commit allows the administrator to specify which claim from
UserInfo should be used as a username; they may, for instance, choose
preferred_username if they know that the values from their OIDC
provider will be unique.  The upshot of this change is that
administrators can now use OIDC providers like Gitea/Forgejo.

Fixes #3008.