frontend: use ORM attributes in queries
See #3130

This is a straightforward case. We cannot join or select using
attribute names represented as strings.

    sqlalchemy.exc.ArgumentError: Strings are not accepted for
    attribute names in loader options; please use class-bound
    attributes directly.

This one is a bit harder to understand.

    sqlalchemy.exc.ArgumentError: expected ORM mapped attribute for
    loader strategy argument

Easier to show on an example. When selecting e.g. `BuildChroot` and
wanting to join the build information, we cannot do
`join(models.Build)` but instead, we need to
`join(models.BuildChroot.build)`.