frontend: automatically EOL lifeless rolling chroots
When rolling policy finishes, we want to remove the data, but we also
want to mark the CoprChroot.deleted (disable it from the project)
because the corresponding mock chroot is still active in Copr.

When user does a new build, we reset the CoprChroot
delete_after/notification state, always.  This is ok because during the
normal MockChroot EOL policy, we do not allow new builds.

Implements: https://github.com/fedora-copr/debate/blob/main/2024-04-23-rawhide-chroots-eol.md
Fixes: #2933