backend: correctly obtain information about a publication
The previous code "worked" only by an accident. Previously we called
the publication endpoint without a trailing slash (changed in d8013c),
which caused some weird redirect and returned a list of all
publications, not the one created publication that I expected.

We need to manually query the publication href from the submitted
task.